+++ src/main/java/com/programmers/jpaboard/controller/UserController.java
@@ -1,8 +1,8 @@
 package com.programmers.jpaboard.controller;
 
-import com.programmers.jpaboard.dto.request.UserCreateRequest;
-import com.programmers.jpaboard.dto.response.ApiResponse;
-import com.programmers.jpaboard.dto.response.UserCreateResponse;
+import com.programmers.jpaboard.dto.user.request.UserCreateRequest;
+import com.programmers.jpaboard.dto.ApiResponse;
+import com.programmers.jpaboard.dto.user.response.UserCreateResponse;
 import com.programmers.jpaboard.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
