+++ src/main/java/com/programmers/jpaboard/dto/post/request/PostUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.programmers.jpaboard.dto.request;
+package com.programmers.jpaboard.dto.post.request;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
