+++ src/main/java/com/programmers/jpaboard/dto/post/response/PostIdResponse.java
@@ -1,4 +1,4 @@
-package com.programmers.jpaboard.dto.response;
+package com.programmers.jpaboard.dto.post.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
