+++ src/main/java/com/programmers/jpaboard/exception/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
 package com.programmers.jpaboard.exception;
 
-import com.programmers.jpaboard.dto.response.ApiResponse;
+import com.programmers.jpaboard.dto.ApiResponse;
 import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
